Update README.md
#Kickstarting with Excel
##Overview of Project
A comparison and overlook of different campaign outcomes in regards to launching dates and funding goals.
###Purpose
##Analysis and Challenges
###Analysis of Outcomes Based on Launch Date
The data shows that the largest amount of successful shows occurred during the Summer. Especially in the months of May, June, and July.
###Analysis of Outcomes Based on Goals
Those with a lower Goal were more successful in comparison to the Kickstarters that had higher goals, especially at 25000 and more.
###Challenges and Difficulties Encountered
When entering the information, correctly entering the data is crucial. One of the beginning issues that I ran into was incorrectly typed data that later also caused incorrect information to be pulled. Working slowly and retracing the steps is what helped in finishing.
##Results
In the Theatre Outcomes Based on Launch Date graph, the failed ad successful lies have the same patterns, however, there is a noticeable peak in the month of October for failed play, showing the month to avoid launching too many plays. The best months would be in the summer which could be connected to the increase in free time many have with school letting out and having the most successful theatre productions. The higher failure productions are those that fall between 45000 to 49999, however, it is also important to note that the high failure [percentage is attributed to there being only one project in that goal. When looking at the date it can be seen that the safest goal point would fall either less than 1000 or 1000 to 4999.
The data that is within the worksheet may need to incorporate the genre or topic of the productions as well, the information is also limited in the fact that it does not state the location of the productions. When adding more information we can look for the most successful production, especially when including graphs that mention the number of backers in comparison to the goal and success.